+++ test/writable.js
@@ -1,42 +1,38 @@
-
 /**
  * Module dependencies.
  */
 
-var Parser = require('../');
-var assert = require('assert');
-var inherits = require('util').inherits;
-var Writable = require('stream').Writable;
-
-// for node v0.6.x-v0.8.x support
-if (!Writable) Writable = require('readable-stream/writable');
+import Parser from '../index.js';
+import assert from 'assert';
+import { inherits } from 'util';
+import { Writable } from 'stream';
 
 describe('Writable streams', function () {
 
-  var val = 1337;
-  var buf = new Buffer(4);
+  const val = 1337;
+  const buf = Buffer.alloc(4);
   buf.writeUInt32LE(val, 0);
 
   it('should have the `_bytes()` function', function () {
-    var w = new Writable();
+    const w = new Writable();
     Parser(w);
     assert.equal('function', typeof w._bytes);
   });
 
   it('should have the `_skipBytes()` function', function () {
-    var w = new Writable();
+    const w = new Writable();
     Parser(w);
     assert.equal('function', typeof w._skipBytes);
   });
 
   it('should *not* have the `_passthrough()` function', function () {
-    var w = new Writable();
+    const w = new Writable();
     Parser(w);
     assert.notEqual('function', typeof w._passthrough);
   });
 
   it('should read 4 bytes in one chunk', function (done) {
-    var w = new Writable();
+    const w = new Writable();
     Parser(w);
 
     // read 4 bytes
@@ -50,7 +46,7 @@ describe('Writable streams', function () {
   });
 
   it('should read 4 bytes in multiple chunks', function (done) {
-    var w = new Writable();
+    const w = new Writable();
     Parser(w);
 
     // read 4 bytes
@@ -60,14 +56,14 @@ describe('Writable streams', function () {
       done();
     });
 
-    for (var i = 0; i < buf.length; i++) {
-      w.write(new Buffer([ buf[i] ]));
+    for (let i = 0; i < buf.length; i++) {
+      w.write(Buffer.from([ buf[i] ]));
     }
     w.end();
   });
 
   it('should read 1 byte, 2 bytes, then 3 bytes', function (done) {
-    var w = new Writable();
+    const w = new Writable();
     Parser(w);
 
     // read 1 byte
@@ -91,7 +87,7 @@ describe('Writable streams', function () {
       done();
     }
 
-    w.end(new Buffer([ 0, 0, 1, 0, 1, 2 ]));
+    w.end(Buffer.from([ 0, 0, 1, 0, 1, 2 ]));
   });
 
   it('should work when mixing in to a subclass\' `prototype`', function (done) {
@@ -104,7 +100,7 @@ describe('Writable streams', function () {
     // mixin to the `prototype`
     Parser(MyWritable.prototype);
 
-    var count = 2;
+    let count = 2;
     MyWritable.prototype.onbytes = function (buf) {
       assert.equal(2, buf.length);
       assert.equal(0, buf[0]);
@@ -113,21 +109,21 @@ describe('Writable streams', function () {
       if (!count) done();
     };
 
-    var a = new MyWritable();
-    var b = new MyWritable();
+    const a = new MyWritable();
+    const b = new MyWritable();
 
     // interleave write()s
-    a.write(new Buffer([ 0 ]));
-    b.write(new Buffer([ 0 ]));
-    a.write(new Buffer([ 1 ]));
-    b.write(new Buffer([ 1 ]));
+    a.write(Buffer.from([ 0 ]));
+    b.write(Buffer.from([ 0 ]));
+    a.write(Buffer.from([ 1 ]));
+    b.write(Buffer.from([ 1 ]));
     a.end();
     b.end();
   });
 
   it('should *not* allow you to buffer Infinity bytes', function () {
     // buffering to Infinity would just be silly...
-    var w = new Writable();
+    const w = new Writable();
     Parser(w);
     assert.throws(function () {
       w._bytes(Infinity);
@@ -135,7 +131,7 @@ describe('Writable streams', function () {
   });
 
   it('should skip 3 bytes then buffer 3 bytes', function (done) {
-    var w = new Writable();
+    const w = new Writable();
     Parser(w);
 
     w._skipBytes(3, function () {
@@ -153,8 +149,8 @@ describe('Writable streams', function () {
   describe('async', function () {
 
     it('should accept a callback function for `_bytes()`', function (done) {
-      var w = new Writable();
-      var data = 'test';
+      const w = new Writable();
+      const data = 'test';
       Parser(w);
       w._bytes(data.length, function (chunk, fn) {
         setTimeout(fn, 25);
@@ -166,7 +162,7 @@ describe('Writable streams', function () {
     });
 
     it('should emit an "error" event when data is written with no parsing function', function (done) {
-      var w = new Writable();
+      const w = new Writable();
       Parser(w);
       w.once('error', function (err) {
         assert(err);
@@ -188,7 +184,7 @@ describe('Writable streams', function () {
     Parser(FrameParser.prototype);
 
     FrameParser.prototype.onsize = function (buf) {
-      var size = buf.readUInt8(0);
+      const size = buf.readUInt8(0);
       this._bytes(size, this.onframe);
     };
 
@@ -200,22 +196,22 @@ describe('Writable streams', function () {
     };
 
     it('should emit 1 "frame" event', function (done) {
-      var p = new FrameParser();
-      var s = 'a string';
+      const p = new FrameParser();
+      const s = 'a string';
       p.on('frame', function (frame) {
         assert.equal(s, frame);
         done();
       });
-      p.write(new Buffer([ s.length ]));
-      p.write(new Buffer(s));
+      p.write(Buffer.from([ s.length ]));
+      p.write(Buffer.from(s));
       p.end();
     });
 
     it('should emit 2 "frame" events', function (done) {
-      var p = new FrameParser();
-      var s = 'a string';
-      var s2 = 'done';
-      var count = 0;
+      const p = new FrameParser();
+      const s = 'a string';
+      const s2 = 'done';
+      let count = 0;
       p.on('frame', function (frame) {
         count++;
         if (s2 == frame) {
@@ -223,10 +219,10 @@ describe('Writable streams', function () {
           done();
         }
       });
-      p.write(new Buffer([ s.length ]));
-      p.write(new Buffer(s));
-      p.write(new Buffer([ s2.length ]));
-      p.write(new Buffer(s2));
+      p.write(Buffer.from([ s.length ]));
+      p.write(Buffer.from(s));
+      p.write(Buffer.from([ s2.length ]));
+      p.write(Buffer.from(s2));
       p.end();
     });